Debugging functionnality
I've added a quite usefull trick to debug running script. It goes as
follow:

1. Import utils.debug, call utils.debug.setdebug() in your code once.

This will configure a signal handler that catches "SIGTSTP" which is
sent to your process by your terminal when you hit "ctrl-Z". Upon
catching the signal, you will be prompted in pdb.

2. Hit "ctrl-Z"

3. Debug.

4. Return to your computation by pressing "c" on the pdb prompt

It might not always works, for example if theano is currently compiling
with gcc, you will probably have problems.