Slight refactor to prevent tight coupling.
mnist_ubyte.py is low-level enough that it seems inappropriate to do
path-munging there, and this also tightly couples it to utils/serial.py
which is undesirable for reusability. I think we should try to keep
low-level stuff as independent as possible, and combine them in higher
level abstractions (e.g. the dataset classes).