Format doc in optimization (partially)
In linear_cg, I put the reference to Wikipedia in the summary section.
"See also" expects another function from the library. When you put
something else there, you get either weird documentation or an error.
The "Notes" section doesn't seem to accept hyperlinks. Finally, in the
Numpy docstring standard, it is discouraged to put web pages in the
"References" section.