Format doc in termination_criteria
The NumPy docstring standard specifies that, in the "Returns" section,
it is not necessary to name the returned value (as long as you mention
its type).

However, the documentation looks a bit strange when doing that. The
variable type is in boldface and is followed by a colon. Is there a way
around this (except giving the return value a name)?