Format doc in utils/data_specs.py (ok?)
I am unsure about the changes here. Using the original version on my
computer, Sphinx was not outputting the "Attributes" sections of class
DataSpecsMapping.

Howewer, it appears in the documentation at deeplearning.net, although
it looks strange. After the small changes I made, I can get Sphinx to
output the doc on my computer, but the format doesn't correspond to that
of the "Parameters" section, nor to that of the "Attributes" section on
deeplearning.net.