Add get_num_examples() method
So far, it is impossible to get the number of examples without getting a
matrix directly which is only possible on dense_design_matrix and
sparse_dataset with get_design_matrix() method.

This commit adds a method get_num_examples() which returns the number
of examples in the dataset. It also deprecates the attribute num_examples
of dense_design_matrix in favor of Dataset.get_num_examples().